Loading moved into layout
